--- 05/src/constants.rs
@@ -1,1 +1,0 @@
-pub const VEC3_LEN: usize = 3;
