+++ 05/src/lib.rs
@@ -1,5 +1,3 @@
-pub mod constants;
 pub mod counters;
-pub mod defaults;
-pub mod types;
+pub mod pair;
 pub mod vectors;
