--- 05/src/types.rs
@@ -1,6 +1,0 @@
-use crate::constants::VEC3_LEN;
-
-pub type Pair = (i32, i32);
-pub type Vec3 = [i32; VEC3_LEN];
-pub type SignedCounter = isize;
-pub type UnsignedCounter = usize;
