+++ 05/src/vectors.rs
@@ -1,18 +1,5 @@
-use crate::constants::VEC3_LEN;
-use crate::defaults::default_vec3;
-use crate::types::Pair;
-use crate::types::Vec3;
-
-#[cfg(test)]
-mod test_vec3_vector_sum {
-    use super::vec3_vector_sum;
-
-    #[test]
-    fn test() {
-        assert_eq!(vec3_vector_sum([1; 3], [2; 3]), [3, 3, 3]);
-        assert_ne!(vec3_vector_sum([1; 3], [1; 3]), [3, 3, 3]);
-    }
-}
+pub const VEC3_LEN: usize = 3;
+pub type Vec3 = [i32; VEC3_LEN];
 
 pub fn vec3_vector_sum(a: Vec3, b: Vec3) -> Vec3 {
     let mut c = default_vec3();
@@ -20,30 +7,6 @@ pub fn vec3_vector_sum(a: Vec3, b: Vec3) -> Vec3 {
         c[i] = a[i] + b[i];
     }
     c
-}
-
-#[cfg(test)]
-mod test_pair_vector_sum {
-    use super::pair_vector_sum;
-
-    #[test]
-    fn test() {
-        assert_eq!(pair_vector_sum((1, 2), (3, 3)), (4, 5));
-    }
-}
-
-pub fn pair_vector_sum(a: Pair, b: Pair) -> Pair {
-    (a.0 + b.0, a.1 + b.1)
-}
-
-#[cfg(test)]
-mod test_vec3_scalar_sum {
-    use super::vec3_scalar_sum;
-
-    #[test]
-    fn test() {
-        assert_eq!(vec3_scalar_sum([1, 2, 3], [1, 2, 3]), 12);
-    }
 }
 
 pub fn vec3_scalar_sum(a: Vec3, b: Vec3) -> i32 {
@@ -54,16 +17,27 @@ pub fn vec3_scalar_sum(a: Vec3, b: Vec3) -> i32 {
     c
 }
 
-#[cfg(test)]
-mod test_pair_scalar_sum {
-    use super::pair_scalar_sum;
-
-    #[test]
-    fn test() {
-        assert_eq!(pair_scalar_sum((1, 1), (2, 2)), 6);
-    }
+pub fn default_vec3() -> Vec3 {
+    [0; 3]
 }
 
-pub fn pair_scalar_sum(a: Pair, b: Pair) -> i32 {
-    a.0 + a.1 + b.0 + b.1
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn vec3_vector_sum_correct_sum_tuple() {
+        assert_eq!(vec3_vector_sum([1; 3], [2; 3]), [3, 3, 3]);
+        assert_ne!(vec3_vector_sum([1; 3], [1; 3]), [3, 3, 3]);
+    }
+
+    #[test]
+    fn vec3_scalar_sum_correct_sum_tuple() {
+        assert_eq!(vec3_scalar_sum([1, 2, 3], [1, 2, 3]), 12);
+    }
+
+    #[test]
+    fn default_vec3_zero_tuple() {
+        assert_eq!(default_vec3(), [0; 3]);
+    }
 }
